Стабилизация системы
